Add files via upload
Cricket is one of the most watched sports in India and IPL is its biggest mode of entertainment. One of the ways the audience gets some insight of the statistics is through various easy to understand visualizations. It helps the BCCI to select players for next domestic as well as international tournaments. This also helps to finally end the constant
debate between the fans on who’s better. To implement this, I’ve created an interactive dashboard with the help of Power BI.
Using Power BI, all the various databases are connected with each other and visualizations are made accordingly and integrated with dashboard for effective statistical analysis.